ValueError breaking the uploading of a folder (#157)
While uploading a folder, files greater than 2GiB is breaking the whole process.🥺

Here, Exception has been caught already, but raising the exception is the main reason of breaking the process.(I guess)

Skip the raise of ValueError may solve this. Because we can't do anything about that limit as it is being imposed by Telegram itself.